Move dashboard/graphics entry points to inside sub-folders
