+++ vite.config.mjs
@@ -16,7 +16,12 @@ export default defineConfig({
     vue({ template: { transformAssetUrls } }),
     quasar({ autoImportComponentCase: 'pascal' }),
     checker({ vueTsc: { tsconfigPath: 'tsconfig.browser.json' } }),
-    NodeCGPlugin(),
+    NodeCGPlugin({
+      inputs: {
+        'graphics/*/main.ts': './src/graphics/template.html',
+        'dashboard/*/main.ts': './src/dashboard/template.html',
+      },
+    }),
   ],
   resolve: {
     alias: {
